Rename HostedFFAAnnouncerTask to HostedGameCountdown.
